better handling of packet SNR, which can be a negative number and is stored in the register as twos complement

+++ src/LoRa.h
@@ -42,7 +42,7 @@ public:
 
   int parsePacket(int size = 0);
   int packetRssi();
-  float packetSnr();
+  int packetSnr();
   long packetFrequencyError();
 
   int rssi();
